Add files via upload
A Pneumonia infection can spread from the lungs into the bloodstream. This is a serious complication. It can reach other major organs and result in organ damage or even death. This risk can be significantly reduced by timely diagnosis and treatment. Systematic Screening has been identified as cost-effective way to save health service resources. 
Automatic Chest X-Ray Image Analysis is emerging as an important screening tool for early Pneumonia detection , which can reduce the workload associated to manual grading as well as save diagnosis cost and time. Many research efforts in the last years has been devoted to developing automated tools to help in the detection and evaluation. 

For this project we picked a dataset of our choice from kaggle and apply the concepts learned to train deep learning models end to end with PyTorch expermenting different hyper parameters and metrics. 

1) Find a dataset online
2) Understand and describe the modeling objective clearly
3) Clean data if required and perform exploratory analysis (plot graphs , ask questions )
4) Modeling
       A) Define a model (Network Architecture)
       B) Pick some hyper parameters
       C) Train the model
       D) Make predictions on samples
       E) Evaluate on test dataset
       F) Save the model weights
       G) Record the metrics
       H) Try different hyper parameters and regularization